refactor: add typescript, eslint and prettier config. remove components to reestructure the files
