--- github-profile-finder/src/App.css
@@ -1,10 +1,0 @@
-#root {
-  width: 100%;
-  min-height: 100vh;
-  font-family: "Open Sans", sans-serif;
-}
-
-#root h2 {
-  text-align: center;
-  margin: 2em 0;
-}
