--- github-profile-finder/src/App.jsx
@@ -1,13 +1,0 @@
-import "./App.css";
-import SearchForm from "./pages/SearchForm";
-
-function App() {
-  return (
-    <div id="root">
-      <h2>GitHub Profile Finder</h2>
-      <SearchForm />
-    </div>
-  );
-}
-
-export default App;
