+++ github-profile-finder/src/App.tsx
@@ -1,0 +1,11 @@
+import { GlobalStyle } from '../styles/GlobalStyles';
+import SearchForm from './pages/SearchForm';
+
+export const App = () => {
+	return (
+		<>
+			<SearchForm />
+			<GlobalStyle />
+		</>
+	);
+};
