--- github-profile-finder/src/components/Advices/Advices.module.css
@@ -1,28 +1,0 @@
-#advice_wrapper {
-  text-align: center;
-  margin: 4em 0;
-}
-
-#advice_wrapper h2 {
-  margin: 1em 0;
-}
-
-#advice_wrapper p {
-  width: 80%;
-  margin: 0 auto;
-  font-size: 1.2em;
-  margin-bottom: 1.2em;
-}
-
-#advice_wrapper button {
-  cursor: pointer;
-  background-color: #aeccda;
-  border-radius: 5px;
-  padding: 0.6em 0.8em;
-  font-weight: bold;
-  border: 1px solid rgb(165, 165, 165);
-}
-
-#advice_wrapper button:hover {
-  opacity: 0.8;
-}
