--- github-profile-finder/src/components/Profile/Profile.module.css
@@ -1,200 +1,0 @@
-#form_wrapper {
-  max-width: 66%;
-  margin: 30px auto 0 auto;
-  min-height: auto;
-  box-shadow: 1px 1px 10px 4px rgba(0, 0, 0, 0.159);
-  border-radius: 10px;
-  padding: 20px;
-  overflow: hidden;
-  background-color: #fcfcfc;
-}
-
-#form_wrapper h3 {
-  font-size: 1.5em;
-  font-weight: bolder;
-  text-align: center;
-  margin-bottom: 1em;
-}
-
-.label {
-  display: flex;
-  flex-direction: column;
-}
-
-.label span {
-  font-weight: bold;
-}
-
-.btn {
-  border: none;
-  background-color: #aeccda;
-  padding: 0.4em 0.8em;
-  border-radius: 5px;
-  font-weight: bolder;
-  cursor: pointer;
-  transition: 0.2s ease-in-out;
-}
-
-.btn:hover {
-  opacity: 0.8;
-}
-
-.error_paragraph {
-  background-color: rgba(252, 149, 149, 0.753);
-  padding: 20px;
-  margin: 1em 0;
-  border-radius: 10px;
-}
-
-input {
-  margin-bottom: 1em;
-  padding: 1em 0;
-  outline: none;
-  border: none;
-  background-color: transparent;
-}
-
-input[type="text"] {
-  border-bottom: 1px solid #ccc;
-}
-
-input[type="submit"] {
-  cursor: pointer;
-  background-color: #aeccda;
-  border-radius: 5px;
-  padding: 0.6em 0.8em;
-  font-weight: bold;
-  border: 1px solid rgb(165, 165, 165);
-  transition: 0.5s ease-in-out;
-}
-
-input[type="submit"]:hover {
-  opacity: 0.8;
-}
-
-input::placeholder {
-  color: #aaa;
-}
-
-#github_profile {
-  padding: 20px;
-  text-align: left;
-}
-
-#github_profile img {
-  background-color: #ccc;
-  width: 260px;
-  height: 260px;
-  border-radius: 50%;
-  display: block;
-  margin: 0 auto;
-  margin-bottom: 2em;
-}
-
-.profile_name {
-  font-size: 24px;
-  line-height: 1.25;
-  display: block;
-  font-weight: bold;
-}
-
-.profile_login {
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 300;
-  line-height: 24px;
-  color: #57606a;
-  margin-bottom: 1em;
-}
-
-.profile_bio {
-  font-size: 16px;
-  margin-bottom: 1em;
-}
-
-.profile_location,
-.profile_blog,
-.profile_link,
-.profile_company {
-  margin-bottom: 0.6em;
-  display: flex;
-  width: 100%;
-  align-items: center;
-  gap: 0.5em;
-}
-
-div p {
-  max-width: 100%;
-  overflow-wrap: break-word;
-  word-wrap: break-word;
-  hyphens: auto;
-}
-
-svg {
-  display: block;
-  min-width: 16px;
-  min-height: 16px;
-}
-
-.profile_blog a {
-  text-decoration: none;
-  color: #000;
-  font-weight: bold;
-}
-
-.profile_url {
-  text-decoration: none;
-  font-weight: bold;
-  color: rgb(0, 0, 0);
-}
-
-.profile_url:hover,
-.profile_blog a:hover {
-  color: #57606a;
-}
-
-.loading_p {
-  padding: 20px;
-  text-align: center;
-  font-size: 1.5em;
-}
-
-@media (max-width: 1024px) {
-  #form_wrapper {
-    max-width: 95%;
-  }
-
-  #form_wrapper h3 {
-    font-size: 20px;
-  }
-
-  #github_profile img {
-    width: 150px;
-    height: 150px;
-  }
-
-  .profile_name {
-    font-size: 20px;
-    line-height: 1.25;
-    display: block;
-    font-weight: bold;
-  }
-
-  .profile_login {
-    font-size: 18px;
-    font-style: normal;
-    font-weight: 300;
-    line-height: 24px;
-    color: #57606a;
-    margin-bottom: 1em;
-  }
-
-  .profile_bio {
-    font-size: 16px;
-    margin-bottom: 1em;
-  }
-
-  svg {
-    font-size: 1em;
-  }
-}
