--- github-profile-finder/src/index.css
@@ -1,11 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  background-color: #f6fcff;
-  width: 100%;
-  min-height: 100vh;
-}
