--- github-profile-finder/src/pages/SearchForm/SearchForm.module.css
@@ -1,4 +1,0 @@
-#search_form_page {
-    width: 100%;
-    height: 100%;
-}
