+++ github-profile-finder/src/pages/SearchForm/index.tsx
@@ -1,0 +1,11 @@
+import { HomePageContainer } from './style';
+
+const SearchForm = () => {
+	return (
+		<HomePageContainer>
+			
+		</HomePageContainer>
+	);
+};
+
+export default SearchForm;
