+++ github-profile-finder/src/pages/SearchForm/style.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const HomePageContainer = styled.div`
+	width: 100%;
+	height: 100%;
+`
